Refactor chat page into modular components
Extract src/app/chat/page.tsx (894 lines) into src/components/chat/:
- types.ts: Type definitions
- message-utils.ts: Pure utility functions for tool extraction and grouping
- use-chat-websocket.ts: Custom hook for WebSocket state management
- message-renderers.tsx: Message type renderers
- tool-renderers.tsx: Tool call/result rendering
- message-groups.tsx: Message grouping and group renderers
- user-message.tsx: User message component
- index.ts: Barrel exports

page.tsx reduced from 894 to 141 lines (84% reduction).
Removes deprecated ToolUseRenderer and ToolResultRenderer.
