+++ src/components/chat/index.ts
@@ -1,0 +1,38 @@
+/**
+ * Chat components barrel exports
+ */
+
+// Components
+export { AssistantGroupRenderer, GroupedMessages } from './message-groups';
+export {
+  AssistantMessageRenderer,
+  ErrorRenderer,
+  MessageRenderer,
+  ResultRenderer,
+  StreamDeltaRenderer,
+  SystemMessageRenderer,
+} from './message-renderers';
+// Utilities
+export {
+  convertHistoryMessage,
+  extractFromContent,
+  extractToolInfo,
+  extractToolResult,
+  extractToolUse,
+  groupMessages,
+  hasToolContent,
+} from './message-utils';
+export { ToolCallGroupRenderer, ToolInfoRenderer } from './tool-renderers';
+// Types
+export type {
+  ChatMessage,
+  ClaudeMessage,
+  HistoryMessage,
+  MessageGroup,
+  ToolInfo,
+  WebSocketMessage,
+} from './types';
+export type { UseChatWebSocketReturn } from './use-chat-websocket';
+// Hooks
+export { useChatWebSocket } from './use-chat-websocket';
+export { UserMessage } from './user-message';
